backend/requirements.txt
# Nova Agent — backend dependencies (direct only; pip resolves the rest).

# ── Web framework ────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart

# ── Database ─────────────────────────────────────────
sqlalchemy
aiosqlite

# ── Config & validation ──────────────────────────────
pydantic
pydantic-settings
python-dotenv

# ── Auth (JWT session tokens; identity via Google OAuth) ──
python-jose[cryptography]

# ── Documents ────────────────────────────────────────
pypdf

# ── Vector store ─────────────────────────────────────
pinecone

# ── LLM / agent framework ────────────────────────────
langgraph
langchain-core
langchain-openai
langchain-google-genai
langchain-ollama
langsmith
openai
google-genai

# ── Google integrations (OAuth + Calendar) ───────────
google-auth-oauthlib
google-api-python-client
